fix the no cameras rendering bug when trying to exit puzzle 7.
previously, the bug caused a no cameras rendering popup to appear when exiting
the stage, either by loading a different scene somehow or pressing the VR back
button on the wall.

turns out the no cameras rendering bug was caused by main menu handler script
component attached to the win menu canvas. puzzle 7 is the only one with the
win menu canvas, so that makes sense. the specific line that seems to trigger
the bug is the PlayMode.overrideMode = true line. not sure why.

also cleaned up a lot of the missing scripts referred to in the PuzzleBase
prefab. i was trying things to fix the bug.